+++ src/core/common/message.cpp
@@ -256,7 +256,7 @@ exit:
     return error;
 }
 
-int Message::Read(uint16_t aOffset, uint16_t aLength, void *aBuf) const
+uint16_t Message::Read(uint16_t aOffset, uint16_t aLength, void *aBuf) const
 {
     Buffer *curBuffer;
     uint16_t bytesCopied = 0;
@@ -622,7 +622,7 @@ MessageQueue::MessageQueue(void)
     mInterface.mTail = NULL;
 }
 
-ThreadError MessageQueue::AddToList(int aList, Message &aMessage)
+ThreadError MessageQueue::AddToList(uint8_t aList, Message &aMessage)
 {
     MessageList *list;
 
@@ -647,7 +647,7 @@ ThreadError MessageQueue::AddToList(int aList, Message &aMessage)
     return kThreadError_None;
 }
 
-ThreadError MessageQueue::RemoveFromList(int aList, Message &aMessage)
+ThreadError MessageQueue::RemoveFromList(uint8_t aList, Message &aMessage)
 {
     MessageList *list;
